requirements.txt
python-dotenv
fastapi
pydantic
uvicorn
pylint
pytest
black